refactor: 모임, 유저 디렉토리 구조 변경으로 인해 깨지는 파일 수정 (#314)
