--- server/momo-api/src/main/java/com/momo/JpaRepositoryScan.java
@@ -1,15 +1,0 @@
-package com.momo;
-
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.FilterType;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-
-@Configuration
-@EnableJpaRepositories(excludeFilters = @ComponentScan.Filter(
-    type = FilterType.ASSIGNABLE_TYPE,
-    classes = {GroupSearchEngine.class})
-)
-public class JpaRepositoryScan {
-
-}
