+++ server/momo-core/src/main/java/com/momo/config/ElasticSearchConfig.java
@@ -1,6 +1,6 @@
 package com.momo.config;
 
-import com.momo.group.domain.repository.GroupSearchEngine;
+import com.momo.group.domain.search.GroupSearchEngine;
 import org.elasticsearch.client.RestHighLevelClient;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
