+++ server/momo-core/src/main/java/com/momo/group/application/GroupService.java
@@ -14,8 +14,8 @@ import com.momo.group.application.dto.GroupSearchConditionRequest;
 import com.momo.group.domain.Group;
 import com.momo.group.domain.participant.Participant;
 import com.momo.group.domain.repository.GroupRepository;
-import com.momo.group.domain.repository.GroupSearchEngine;
 import com.momo.group.domain.repository.ParticipantRepository;
+import com.momo.group.domain.search.GroupSearchEngine;
 import com.momo.user.domain.User;
 import com.momo.user.domain.repository.UserRepository;
 import java.util.List;
