+++ server/momo-core/src/main/java/com/momo/group/domain/repository/GroupRepositoryCustomImpl.java
@@ -1,4 +1,4 @@
-package com.momo.group.infrastructure;
+package com.momo.group.domain.repository;
 
 import static com.momo.achievementrate.entity.QGroupAchievementRate.groupAchievementRate;
 import static com.momo.favorite.entity.QFavoriteGroup.favoriteGroup;
