+++ server/momo-core/src/test/java/com/momo/unit/group/service/GroupServiceTest.java
@@ -28,8 +28,8 @@ import com.momo.group.application.dto.GroupSearchConditionRequest;
 import com.momo.group.domain.Group;
 import com.momo.group.domain.participant.Participant;
 import com.momo.group.domain.repository.GroupRepository;
-import com.momo.group.domain.repository.GroupSearchEngine;
 import com.momo.group.domain.repository.ParticipantRepository;
+import com.momo.group.domain.search.GroupSearchEngine;
 import com.momo.user.domain.User;
 import com.momo.user.domain.repository.UserRepository;
 import java.util.List;
