+++ server/momo-socket/src/main/java/com/momo/JpaRepositoryScan.java
@@ -1,5 +1,6 @@
 package com.momo;
 
+import com.momo.group.domain.search.GroupSearchEngine;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.FilterType;
